Renamed dbOperations.ts file
